+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## Master
 
--
+## 0.1.1 (2020-09-28)
+
+- Add Stripe OAuth flow
+- Add Payment Gateway model
 
 ## 0.1.0 (2020-09-02)